refactor y colocar correctamente los archivos en sus carpetas
