--- api/src/main/java/com/grupo4/todolist/api/Model/JpaRepository.java
@@ -1,5 +1,0 @@
-package com.grupo4.todolist.api.Model;
-
-public interface JpaRepository<T1, T2> {
-
-}
